perfect, but now I want you to integrate a chatbot UI my chatbot which I designed to response to question regarding my professional background, here is the api to post query: https://amaanp314-portfolio-chatbot-api.hf.space/chat
here you can send it a post request like this:

const res = await fetch(apiUrl, { method: "POST", headers: { "Content-Type": "application/json" }, body: JSON.stringify({ message, session_id: sessionId }), }); const data = await res.json();
here, in the session_id, generate any random id

the response you will get in like this data.response

Also the response is going to be in markdown format, and could sometime also consist of code in markdown format, so handle the response accordingly.

Design:

I want a circular icon on the bottom-left of our website.
When we click on it, a input box should come out of it.
The chatbot section should cover 1/3 of the page horizontally
The background of the chatbot UI section should be transparent, so that we could see behind